Code clean up part 3 (Last)
