+++ src/main/java/space/hideaway/controllers/DashboardController.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import space.hideaway.model.User;
-import space.hideaway.services.UserServiceImplementation;
+import space.hideaway.services.user.UserServiceImplementation;
 
 
 /**
