+++ src/main/java/space/hideaway/services/site/SiteStatisticsServiceImplementation.java
@@ -39,9 +39,12 @@ public class SiteStatisticsServiceImplementation implements SiteStatisticsServic
     }
 
     /**
+     * Calculates the statistics for a site. This should be called after new
+     * records have been added to the site. Statistics are calculated from a
+     * pool of available threads, so not to slow down the server.
      *
-     * @param site
-     * @return
+     * @param site The site to have statistics recalculated
+     * @return The recalculated statistics
      */
     @Override
     public Future<SiteStatistics> recalculateStatistics(Site site)
@@ -50,12 +53,25 @@ public class SiteStatisticsServiceImplementation implements SiteStatisticsServic
         return executorService.submit(new StatisticsCalculationRunnable(site));
     }
 
+    /**
+     * Saves site statistics into the database.
+     *
+     * @param siteStatistics The site statistics to save
+     * @return The saved site statistics
+     */
     @Override
     public SiteStatistics save(SiteStatistics siteStatistics)
     {
         return siteStatisticsRepository.save(siteStatistics);
     }
 
+    /**
+     * Gets the most recent statistics for a site. If the site has no statistics
+     * an 'empty' siteStatistics object is return with all statistics set to zero.
+     *
+     * @param site The site that statistics are being requested
+     * @return The statistics
+     */
     @Override
     public SiteStatistics getMostRecent(Site site)
     {
@@ -69,6 +85,9 @@ public class SiteStatisticsServiceImplementation implements SiteStatisticsServic
         }
     }
 
+    /**
+     * A runnable class to calculate the statistics of a site.
+     */
     private class StatisticsCalculationRunnable implements Callable<SiteStatistics>
     {
 
@@ -83,15 +102,25 @@ public class SiteStatisticsServiceImplementation implements SiteStatisticsServic
             this.site = site;
         }
 
+        /**
+         * The starting point on this class. Injects data sets in to
+         * a summary statistics object for calculations.
+         *
+         * @return The recalculated statistics
+         * @throws Exception
+         */
         @Override
         public SiteStatistics call() throws Exception
         {
+            // Gets data for a period of time
             weekData = dataService.getHistoric(HistoryUnit.WEEK, site);
             monthData = dataService.getHistoric(HistoryUnit.LAST_30, site);
             yearData = dataService.getHistoric(HistoryUnit.YEAR, site);
             allData = dataService.getHistoric(HistoryUnit.ALL, site);
 
             SiteStatistics siteStatistics = new SiteStatistics();
+
+            // A class containing common statistical information about a set of values. (min, max, avg...)
             SummaryStatistics summaryStatistics = new SummaryStatistics();
 
             SummaryStatistics weekStatistics = populateStatistics(summaryStatistics, weekData);
@@ -114,6 +143,13 @@ public class SiteStatisticsServiceImplementation implements SiteStatisticsServic
             return siteStatistics;
         }
 
+        /**
+         * Handles the process of injecting data into the summary statistics class for calculation.
+         *
+         * @param summaryStatistics The summary statistics object
+         * @param someData The data set to calculate statistics with
+         * @return The summary statistics object
+         */
         private SummaryStatistics populateStatistics(SummaryStatistics summaryStatistics, List<Data> someData)
         {
             summaryStatistics.clear();
