+++ src/main/java/space/hideaway/util/temperature/TemperatureUnit.java
@@ -1,4 +1,4 @@
-package space.hideaway.util;
+package space.hideaway.util.temperature;
 
 /**
  * Created by dough on 1/24/2017.
