+++ src/main/java/space/hideaway/validation/UserAccountValidator.java
@@ -7,9 +7,11 @@ import org.springframework.validation.ValidationUtils;
 import org.springframework.validation.Validator;
 import space.hideaway.exceptions.UserNotFoundException;
 import space.hideaway.model.User;
-import space.hideaway.services.UserService;
+import space.hideaway.services.user.UserService;
 
-
+/**
+ * Validates a new account.
+ */
 @Component
 public class UserAccountValidator implements Validator
 {
@@ -28,12 +30,18 @@ public class UserAccountValidator implements Validator
         return User.class.equals(aClass);
     }
 
-
+    /**
+     * Validates a new account. Checks for a correct email, username, and password.
+     *
+     * @param o the new user
+     * @param errors the bindingResult
+     */
     @Override
     public void validate(Object o, Errors errors)
     {
         User user = (User) o;
 
+        // Email is not empty or a duplicate.
         ValidationUtils.rejectIfEmptyOrWhitespace(errors, "email", "NotEmpty");
         User byEmail = userService.findByEmail(user.getEmail());
         if (byEmail != null)
@@ -41,8 +49,8 @@ public class UserAccountValidator implements Validator
             errors.rejectValue("email", "Duplicate.userForm.user.email");
         }
 
+        // Username is between 6 and 27, or a duplicate.
         ValidationUtils.rejectIfEmptyOrWhitespace(errors, "username", "NotEmpty");
-
         if (user.getUsername().length() < 6 || user.getUsername().length() > 32)
         {
             errors.rejectValue("username", "Size.userForm.username");
@@ -58,12 +66,12 @@ public class UserAccountValidator implements Validator
 
         }
 
+        // Password is between 8 and 32, or matches.
         ValidationUtils.rejectIfEmptyOrWhitespace(errors, "password", "NotEmpty");
         if (user.getPassword().length() < 8 || user.getPassword().length() > 32)
         {
             errors.rejectValue("password", "Size.userForm.password");
         }
-
         if (!user.getConfirmationPassword().equals(user.getPassword()))
         {
             errors.rejectValue("confirmationPassword", "Diff.userForm.passwordConfirm");
