+++ src/main/resources/templates/station.html
@@ -4,15 +4,14 @@
     <title>Site</title>
     <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous"/>
-
     <link type="text/css" rel="stylesheet" href="https://cdn.datatables.net/1.10.12/css/dataTables.bootstrap.min.css"/>
     <link href="//cdn.datatables.net/responsive/2.1.0/css/dataTables.responsive.css"/>
     <link rel="stylesheet"
           href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-select/1.11.2/css/bootstrap-select.min.css"/>
-    <link type="text/css" rel="stylesheet" th:href="@{/sass/global.css}"/>
-    <link rel="stylesheet" th:href="@{/sass/station.css}"/>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/latest/toastr.min.css"/>
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.0.3/dist/leaflet.css"/>
+    <link type="text/css" rel="stylesheet" th:href="@{/sass/global.css}"/>
+    <link type="text/css" rel="stylesheet" th:href="@{/sass/station.css}"/>
 
 </head>
 <body>
@@ -149,9 +148,7 @@
 </div>
 
 <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-
-<script
-        src="//code.jquery.com/jquery-3.1.1.min.js"
+<script src="//code.jquery.com/jquery-3.1.1.min.js"
         integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="
         crossorigin="anonymous"></script>
 <script src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
